# api/internal/server/download.go
package server

import (
	"database/sql"
	"errors"
	"log"
	"mime"
	"net/http"
	"path/filepath"
	"strconv"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/jackc/pgx/v5"
	"github.com/wenealves10/yt-dlp-downloader/internal/db"
	"github.com/wenealves10/yt-dlp-downloader/internal/media"
	"github.com/wenealves10/yt-dlp-downloader/internal/utils"
)

// Types for download requests and responses

type DownloadType string

const (
	DownloadTypeVideo DownloadType = "video"
	DownloadTypeMusic DownloadType = "music"
	downloadURLTTL                 = time.Minute
)

type downloadRequest struct {
	Type DownloadType `json:"type"  binding:"required,oneof=video music"`
	URL  string       `json:"url"  binding:"required,url,valid_url"`
}

type downloadResponse struct {
	ID              string                `json:"id"`
	Title           string                `json:"title"`
	OriginalUrl     string                `json:"original_url"`
	Format          db.CoreFormatType     `json:"format"`
	Status          db.CoreDownloadStatus `json:"status"`
	ThumbnailUrl    string                `json:"thumbnail_url,omitempty"`
	FileUrl         string                `json:"file_url,omitempty"`
	ExpiresAt       string                `json:"expires_at,omitempty"`
	DurationSeconds int32                 `json:"duration_seconds,omitempty"`
	// ErrorMessage é a mensagem PÚBLICA. O motivo técnico vai em ErrorDetail e
	// só é preenchido para o super admin.
	ErrorMessage  string `json:"error_message,omitempty"`
	ErrorDetail   string `json:"error_detail,omitempty"`
	CreatedAt     string `json:"created_at"`
	Platform      string `json:"platform,omitempty"`
	PlatformLabel string `json:"platform_label,omitempty"`
	// Provider é o nome do mecanismo interno ("yt-dlp"). Também só sai para o
	// super admin: o cliente final não tem por que saber com o que baixamos.
	Provider      string `json:"provider,omitempty"`
	QualityLabel  string `json:"quality_label,omitempty"`
	Uploader      string `json:"uploader,omitempty"`
	FileSizeBytes int64  `json:"file_size_bytes,omitempty"`
}

// formatarInstante protege contra o ponteiro nulo do modelo gerado.
func formatarInstante(instante *time.Time) string {
	if instante == nil {
		return ""
	}
	return instante.Format(time.RFC3339)
}

// montarDownloadResponse traduz a linha do banco para a resposta da API.
//
// O parâmetro `operador` é o que separa as duas visões, e está numa função só
// para não haver dois lugares decidindo o que pode sair: o cliente final recebe
// a mensagem pública e nada mais; o super admin recebe também o motivo técnico
// e o nome do mecanismo interno, que é o que permite diagnosticar sem entrar no
// container.
func montarDownloadResponse(download db.Download, operador bool) downloadResponse {
	item := downloadResponse{
		ID:              download.ID.String(),
		Title:           download.Title,
		Status:          download.Status,
		OriginalUrl:     download.OriginalUrl,
		ThumbnailUrl:    download.ThumbnailUrl.String,
		FileUrl:         download.FileUrl.String,
		DurationSeconds: download.DurationSeconds.Int32,
		ExpiresAt:       download.ExpiresAt.Time.Format(time.RFC3339),
		ErrorMessage:    download.ErrorMessage.String,
		Format:          download.Format,
		// CreatedAt é ponteiro no modelo gerado. A coluna é NOT NULL, mas
		// desreferenciar sem guarda transforma uma linha inesperada em pânico
		// que derruba a listagem inteira — e não só aquele item.
		CreatedAt:     formatarInstante(download.CreatedAt),
		Platform:      download.Platform,
		PlatformLabel: media.Platform(download.Platform).Label(),
		QualityLabel:  download.QualityLabel.String,
		Uploader:      download.Uploader.String,
		FileSizeBytes: download.FileSizeBytes,
	}

	if operador {
		item.ErrorDetail = download.ErrorDetail.String
		item.Provider = download.Provider.String
	}
	return item
}

// createDownload é o endpoint antigo, mantido para clientes que já existiam.
// Ele delega ao mesmo caminho do downloader multiplataforma: manter uma segunda
// implementação de download aqui significaria corrigir cada bug duas vezes.
//
// O contrato de entrada continua o mesmo ({type, url}), e a validação de URL
// segue restrita ao YouTube, como sempre foi para esta rota.
func (s *Server) createDownload(ctx *gin.Context) {
	var req downloadRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request format"})
		return
	}

	kind := string(media.KindVideo)
	if req.Type == DownloadTypeMusic {
		kind = string(media.KindAudio)
	}

	// Sem format_id: o provider escolhe o melhor formato do tipo pedido, que é
	// exatamente o que esta rota sempre fez.
	s.criarDownloadMedia(ctx, criarDownloadMediaRequest{URL: req.URL, Kind: kind})
}

func (s *Server) getDownloads(ctx *gin.Context) {
	userID, autenticado := usuarioAtualID(ctx)
	if !autenticado {
		return
	}

	pageStr := ctx.DefaultQuery("page", "1")
	perPageStr := ctx.DefaultQuery("perPage", "10")

	page, err := strconv.Atoi(pageStr)
	if err != nil || page < 1 {
		page = 1
	}
	perPage, err := strconv.Atoi(perPageStr)
	if err != nil || perPage < 1 {
		perPage = 10
	}

	offset := (page - 1) * perPage

	totalDownloads, err := s.store.CountDownloadsByUser(ctx, userID)
	if err != nil {
		log.Printf("Failed to count downloads: %v", err)
		ctx.JSON(500, gin.H{"error": "Failed to count downloads"})
		return
	}

	downloads, err := s.store.GetDownloadsByUser(ctx, db.GetDownloadsByUserParams{
		UserID: userID,
		Limit:  int32(perPage),
		Offset: int32(offset),
	})
	if err != nil {
		log.Printf("Failed to get downloads: %v", err)
		ctx.JSON(500, gin.H{"error": "Failed to get downloads"})
		return
	}

	// O detalhe técnico e o nome do provider só saem para quem opera.
	operador := false
	if user, existe := currentUser(ctx); existe {
		operador = user.Role == db.CoreUserRoleSuperAdmin
	}

	var response []downloadResponse
	for _, download := range downloads {
		response = append(response, montarDownloadResponse(download, operador))
	}

	ctx.JSON(http.StatusOK, gin.H{
		"next_page": (page * perPage) < int(totalDownloads),
		"prev_page": page > 1,
		"page":      page,
		"per_page":  perPage,
		"total":     totalDownloads,
		"downloads": response,
	})
}

func (s *Server) getDailyDownloads(ctx *gin.Context) {
	user, autenticado := usuarioAtual(ctx)
	if !autenticado {
		ctx.JSON(http.StatusUnauthorized, gin.H{"error": "requisição não autenticada"})
		return
	}
	userID := user.ID

	dailyDownloads, err := s.store.CountDownloadsToday(ctx, userID)
	if err != nil {
		log.Printf("Failed to count daily downloads: %v", err)
		ctx.JSON(500, gin.H{"error": "Failed to count daily downloads"})
		return
	}
	// O super admin não tem teto; o contador precisa dizer isso, ou a tela
	// bloqueia o campo de URL assim que o limite nominal é atingido.
	if user.Role == db.CoreUserRoleSuperAdmin {
		ctx.JSON(http.StatusOK, gin.H{
			"daily_downloads": dailyDownloads,
			"daily_limit":     0,
			"remaining":       0,
			"unlimited":       true,
		})
		return
	}

	ctx.JSON(http.StatusOK, gin.H{
		"daily_downloads": dailyDownloads,
		"daily_limit":     user.DailyLimit,
		"remaining":       int64(user.DailyLimit) - dailyDownloads,
		"unlimited":       false,
	})
}

// authorizedCompletedDownload loads a download after checking both ownership
// and availability. It keeps all download delivery paths from ever accepting
// an object key supplied by the browser.
func (s *Server) authorizedCompletedDownload(ctx *gin.Context) (db.Download, bool) {
	userID, autenticado := usuarioAtualID(ctx)
	if !autenticado {
		return db.Download{}, false
	}

	downloadID, err := utils.ParseUUID(ctx.Param("id"))
	if err != nil {
		erroDownload(ctx, http.StatusBadRequest, CodeInvalidRequest, "identificador de download inválido")
		return db.Download{}, false
	}

	download, err := s.store.GetDownloadByID(ctx.Request.Context(), downloadID)
	if err != nil {
		if errors.Is(err, pgx.ErrNoRows) || errors.Is(err, sql.ErrNoRows) {
			erroDownload(ctx, http.StatusNotFound, CodeNotFound, "download não encontrado")
			return db.Download{}, false
		}
		log.Printf("Failed to get download %s: %v", downloadID, err)
		erroDownload(ctx, http.StatusInternalServerError, CodeInternalError, "falha ao carregar o download")
		return db.Download{}, false
	}

	// Download de outro dono responde 404, e não 403.
	//
	// Um 403 confirmaria que aquele identificador existe. Na tela isso é
	// irrelevante — o usuário só vê os próprios —, mas esta mesma função atende
	// a API de integrações, onde vários sistemas convivem: ali, a diferença
	// entre "não existe" e "existe e não é seu" é um oráculo para enumerar os
	// downloads dos outros.
	if download.UserID != userID {
		erroDownload(ctx, http.StatusNotFound, CodeNotFound, "download não encontrado")
		return db.Download{}, false
	}

	if download.Status != db.CoreDownloadStatusCOMPLETED || !download.FileUrl.Valid || download.FileUrl.String == "" {
		erroDownload(ctx, http.StatusConflict, CodeDownloadNotReady, "o download ainda não está pronto")
		return db.Download{}, false
	}

	if download.ExpiresAt.Valid && !download.ExpiresAt.Time.After(time.Now()) {
		erroDownload(ctx, http.StatusGone, CodeDownloadExpired, "o arquivo deste download expirou")
		return db.Download{}, false
	}

	return download, true
}

// erroDownload responde um erro das rotas de download com mensagem E código.
//
// Estas rotas atendem as duas superfícies, e o código é obrigatório na de
// integrações: é por ele que o sistema cliente decide entre tentar de novo mais
// tarde (download_not_ready) e desistir de vez (download_expired).
func erroDownload(ctx *gin.Context, status int, codigo, mensagem string) {
	registrarCodigoErro(ctx, codigo)
	ctx.JSON(status, gin.H{"error": mensagem, "code": codigo})
}

// downloadURL first authenticates the account in the API and only then creates
// a one-minute signed R2 URL. The URL is never stored, returned in the history
// payload, or derived from the browser-provided object key.
func (s *Server) downloadURL(ctx *gin.Context) {
	download, ok := s.authorizedCompletedDownload(ctx)
	if !ok {
		return
	}

	contentType := mime.TypeByExtension("." + strings.ToLower(string(download.Format)))
	if contentType == "" {
		contentType = "application/octet-stream"
	}

	filename := downloadFilename(download.Title, string(download.Format))
	presignedURL, err := s.storage.PresignDownload(
		ctx.Request.Context(),
		download.FileUrl.String,
		filename,
		contentType,
		downloadURLTTL,
	)
	if err != nil {
		log.Printf("Failed to create R2 download URL for %s: %v", download.ID, err)
		ctx.JSON(http.StatusBadGateway, gin.H{"error": "Failed to prepare download"})
		return
	}

	ctx.Header("Cache-Control", "no-store")
	ctx.JSON(http.StatusOK, gin.H{
		"url":        presignedURL,
		"expires_at": time.Now().Add(downloadURLTTL).UTC().Format(time.RFC3339),
	})
}

// downloadFile is retained for authenticated API clients that need the API to
// stream bytes. The web UI uses downloadURL instead, which keeps large file
// transfers in R2 while still requiring an API authorization check first.
func (s *Server) downloadFile(ctx *gin.Context) {
	download, ok := s.authorizedCompletedDownload(ctx)
	if !ok {
		return
	}

	file, err := s.storage.OpenFile(ctx.Request.Context(), download.FileUrl.String)
	if err != nil {
		log.Printf("Failed to open download %s from storage: %v", download.ID, err)
		ctx.JSON(http.StatusBadGateway, gin.H{"error": "Failed to retrieve download file"})
		return
	}
	defer file.Body.Close()

	contentType := file.ContentType
	if contentType == "" {
		contentType = mime.TypeByExtension("." + strings.ToLower(string(download.Format)))
	}
	if contentType == "" {
		contentType = "application/octet-stream"
	}

	filename := downloadFilename(download.Title, string(download.Format))
	ctx.DataFromReader(http.StatusOK, file.ContentLength, contentType, file.Body, map[string]string{
		"Cache-Control":          "no-store",
		"Content-Disposition":    mime.FormatMediaType("attachment", map[string]string{"filename": filename}),
		"X-Content-Type-Options": "nosniff",
	})
}

func downloadFilename(title, format string) string {
	filename := strings.TrimSpace(title)
	filename = strings.Map(func(r rune) rune {
		switch r {
		case '/', '\\', '\x00', '\r', '\n':
			return '_'
		default:
			return r
		}
	}, filename)
	if filename == "" || filename == "." {
		filename = "download"
	}

	extension := "." + strings.ToLower(format)
	if filepath.Ext(filename) == "" && extension != "." {
		filename += extension
	}
	return filename
}

func (s *Server) deleteDownload(ctx *gin.Context) {
	userID, autenticado := usuarioAtualID(ctx)
	if !autenticado {
		return
	}

	downloadID := ctx.Param("id")
	if downloadID == "" {
		ctx.JSON(400, gin.H{"error": "Download ID is required"})
		return
	}

	downloadIDParsed, err := utils.ParseUUID(downloadID)
	if err != nil {
		log.Printf("Failed to parse download ID: %v", err)
		ctx.JSON(400, gin.H{"error": "Invalid download ID"})
		return
	}

	downloadExists, err := s.store.GetDownloadByID(ctx, downloadIDParsed)
	if err != nil {
		log.Printf("Not found download: %v\n", err)
		ctx.JSON(404, gin.H{"error": "Download not found"})
		return
	}

	if downloadExists.UserID != userID {
		log.Printf("Unauthorized access to download: %s by user: %s", downloadID, userID)
		ctx.JSON(403, gin.H{"error": "You do not have permission to delete this download"})
		return
	}

	if err := s.store.DeleteDownload(ctx, downloadIDParsed); err != nil {
		log.Printf("Failed to delete download: %v", err)
		ctx.JSON(500, gin.H{"error": "Failed to delete download"})
		return
	}

	if err := s.storage.DeleteFile(ctx, downloadExists.FileUrl.String); err != nil {
		log.Printf("Failed to delete file from storage: %v", err)
	}

	ctx.JSON(http.StatusNoContent, gin.H{})
}